+++ src/EventSourcingDB/EventSourcingDB.js
@@ -124,7 +124,7 @@ class EventSourcingDB extends EventEmitter {
 		withViews = true,
 		onWillOpen,
 		onBeforeMigrations: prevOBM,
-		onDidOpen: prevODO,
+		onDidOpen,
 		...dbOptions
 	}) {
 		super()
@@ -140,20 +140,11 @@ class EventSourcingDB extends EventEmitter {
 			...dbOptions,
 			onWillOpen,
 			onBeforeMigrations: async db => {
-				// hacky side-channel to get current version to queue without deadlocks
-				// TODO pass this immediately and let queue handle it itself on open
-				this._knownV = await db.userVersion()
+				const v = await db.userVersion()
+				if (v) this.queue.setKnownV(v)
 				if (prevOBM) await prevOBM()
 			},
-			onDidOpen: async db => {
-				// let's hope nobody added events to the queue with the wrong version
-				const {_knownV} = this
-				if (_knownV) {
-					this._knownV = null
-					await this.queue.setKnownV(_knownV)
-				}
-				if (prevODO) await prevODO(db)
-			},
+			onDidOpen,
 		})
 		const {readOnly} = this.rwDb
 
@@ -178,14 +169,6 @@ class EventSourcingDB extends EventEmitter {
 				...dbOptions,
 				name: `${dbOptions.name || ''}Queue`,
 				file: queueFile || this.rwDb.file,
-				onDidOpen: async () => {
-					// let's hope nobody added events to the queue with the wrong version
-					const {_knownV} = this
-					if (_knownV) {
-						this._knownV = null
-						await this.queue.setKnownV(_knownV)
-					}
-				},
 			})
 			this.queue = new EventQueue({
 				db: qDb,
@@ -404,12 +387,7 @@ class EventSourcingDB extends EventEmitter {
 	}
 
 	async dispatch(type, data, ts) {
-		const {_knownV, queue} = this
-		if (_knownV) {
-			this._knownV = null
-			await queue.setKnownV(_knownV)
-		}
-		const event = await queue.add(type, data, ts)
+		const event = await this.queue.add(type, data, ts)
 		return this.handledVersion(event.v)
 	}