+++ Changelog.md
@@ -5,6 +5,11 @@
 ### Changes
 
 - `sqlite.userVersion(v?: number)` now always returns the user version, also when setting it. This makes more sense from an API perspective and has as good as no performance impact.
+- EventQueue: `.setKnownV()` is now synchronous and no longer returns a Promise. This is only a breaking change if you were using `.then()` instead of `await`.
+
+### Fixes
+
+- ESDB: fix deadlock on queue add in migration (due to version setting)
 
 ## 3.3.2