+++ Changelog.md
@@ -8,6 +8,7 @@
 - EventQueue: `.setKnownV()` is now synchronous and no longer returns a Promise. This is only a breaking change if you were using `.then()` instead of `await`.
 - ESDB: Added `transact({event, model, store, dispatch})` phase to the event processing flow. In this callback, you can call `dispatch` to generate and await sub-events, and calling ESModel will work too (any model can use the `dispatch` given via the constructor).
   This requires the use of `AsyncLocalStorage`, and thus the minimum NodeJS version is now v12.17
+- ESDB: `dispatch({type, data, ts})` (passing everything in a single argument) is now also possible, as well as for `addEvent`.
 
 ### Fixes